+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Replace ghostwheel with guardrails
 - Ensure index-io is present to compute completions
 - Add flag to enable/disable trace support on query editor
+- Query editor supports sync and async parsers
 
 ## [1.0.7]
 - Fix graph comm in attribute graph panel